+++ include/neuron.h
@@ -4,17 +4,17 @@ class Neuron{
     public:
         Neuron(int no_inputs, bool last_layer);
 
-        static constexpr float LEAKY_RELU_ALPHA = 0.1;
-        static constexpr float LEARNING_RATE = 0.001;
+        static constexpr double LEAKY_RELU_ALPHA = 0.1;
+        static constexpr double LEARNING_RATE = 0.001;
 
-        float signal_error;
+        double signal_error;
 
         bool last_layer_;
-        float bias_;
-        std::vector<float> weights_;
-        std::vector<float> inputs_;
+        double bias_;
+        std::vector<double> weights_;
+        std::vector<double> inputs_;
 
-        float calculate_neuron_output();
+        double calculate_neuron_output();
 
         void calculate_signal_error();
 };
